+++ client/faculdade-ead/src/app/(authenticated)/Aluno/Head.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Head = () => {
+  return (
+    <>
+      <title>Área do aluno - Faculdade EAD</title>
+    </>
+  );
+};
+
+export default Head;
