+++ client/faculdade-ead/src/app/(authenticated)/Aluno/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Aluno = () => {
+  return <div>Aluno</div>;
+};
+
+export default Aluno;
