+++ client/faculdade-ead/src/public/_mixin.sass
@@ -1,12 +1,14 @@
 @use './variables' as global
 
 @mixin btn
+  align-items: center
   background-color: global.$bg-btn
   font-size: global.$font-size
   color: global.$font-color
-  width: 222.36px
-  height: 50px
+  width: 200px
+  height: 40px
   border-radius: 20px
-  margin-right: 4px
+  margin-right: 12px
+  margin-top: 5px
   border: none
   padding: 0
