+++ client/faculdade-ead/src/public/variables.sass
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,400;0,500;0,600;0,800;0,900;1,400;1,500;1,600&display=swap')
+
 $font-family: 'Montserrat', sans-serif
 $font-size: 18px
 $font-color: #163020
@@ -5,5 +7,5 @@ $font-color-alt: #FAF8ED
 $bg-color: #FAF8ED
 $bg-btn: #D4E7C5
 $bg-btn-alt: #163020
-$bg-cards: #618264
+$bg-cards: #79AC78
 $arrow-color: #F9B572
